Move `collect_crate_types` to `rustc_interface`, and use new attribute parser
