+++ tests/ui/attributes/crate-type-macro-call.stderr
@@ -1,25 +1,8 @@
-error: cannot find macro `foo` in the current scope when looking from the crate root
+error: attribute value must be a literal
   --> $DIR/crate-type-macro-call.rs:1:17
    |
 LL | #![crate_type = foo!()]
-   |                 ^^^ not found from the crate root
-   |
-   = warning: this was previously accepted by the compiler but is being phased out; it will become a hard error in a future release!
-   = note: for more information, see issue #124535 <https://github.com/rust-lang/rust/issues/124535>
-   = help: import `macro_rules` with `use` to make it callable above its definition
-   = note: `#[deny(out_of_scope_macro_calls)]` (part of `#[deny(future_incompatible)]`) on by default
+   |                 ^^^^^^
 
 error: aborting due to 1 previous error
-
-Future incompatibility report: Future breakage diagnostic:
-error: cannot find macro `foo` in the current scope when looking from the crate root
-  --> $DIR/crate-type-macro-call.rs:1:17
-   |
-LL | #![crate_type = foo!()]
-   |                 ^^^ not found from the crate root
-   |
-   = warning: this was previously accepted by the compiler but is being phased out; it will become a hard error in a future release!
-   = note: for more information, see issue #124535 <https://github.com/rust-lang/rust/issues/124535>
-   = help: import `macro_rules` with `use` to make it callable above its definition
-   = note: `#[deny(out_of_scope_macro_calls)]` (part of `#[deny(future_incompatible)]`) on by default
